[CONTROLLER] Setup API Gateaway for controller and integrate it

+++ src/main/java/id/ac/ui/cs/advprog/heymartstore/controller/ProductController.java
@@ -5,8 +5,11 @@ import id.ac.ui.cs.advprog.heymartstore.dto.DeleteProductRequest;
 import id.ac.ui.cs.advprog.heymartstore.dto.ModifyProductResponse;
 import id.ac.ui.cs.advprog.heymartstore.model.Product;
 import id.ac.ui.cs.advprog.heymartstore.model.ProductBuilder;
+import id.ac.ui.cs.advprog.heymartstore.service.JwtService;
 import id.ac.ui.cs.advprog.heymartstore.service.ProductService;
+import lombok.RequiredArgsConstructor;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.beans.factory.annotation.Value;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.HttpStatusCode;
 import org.springframework.http.ResponseEntity;
@@ -17,10 +20,15 @@ import java.util.Optional;
 
 @RequestMapping("/product")
 @RestController
+@RequiredArgsConstructor
 public class ProductController {
+    @Value("${spring.route.gateway_url}")
+    private String GATEWAY_URL;
 
     @Autowired
     private ProductService productService;
+
+    private final JwtService jwtService;
 
     @GetMapping("/all-product")
     public ResponseEntity<List<Product>> allProduct() {
@@ -35,7 +43,15 @@ public class ProductController {
     }
 
     @PostMapping("/create")
-    public ResponseEntity<Product> createProduct(@RequestBody CreateProductRequest product) {
+    public ResponseEntity<Product> createProduct(@RequestHeader (value = "Authorization") String id,
+                                                 @RequestBody CreateProductRequest product) throws IllegalAccessException {
+
+        String token = id.replace("Bearer ", "");
+        if (!jwtService.extractRole(token).equalsIgnoreCase("manager")
+                && !jwtService.extractRole(token).equalsIgnoreCase("admin")) {
+            throw new IllegalAccessException("You have no access.");
+        }
+
         Product savedProduct = productService.createProduct(new ProductBuilder()
                 .setName(product.name)
                 .setStock(product.stock)
@@ -45,7 +61,14 @@ public class ProductController {
     }
 
     @PostMapping("/edit")
-    public ResponseEntity<Product> editProduct(@RequestBody ModifyProductResponse product) {
+    public ResponseEntity<Product> editProduct(@RequestHeader (value = "Authorization") String id,
+                                               @RequestBody ModifyProductResponse product) throws IllegalAccessException {
+        String token = id.replace("Bearer ", "");
+        if (!jwtService.extractRole(token).equalsIgnoreCase("manager")
+                && !jwtService.extractRole(token).equalsIgnoreCase("admin")) {
+            throw new IllegalAccessException("You have no access.");
+        }
+
         Product savedProduct = productService.editProduct(product.UUID,
                 new ProductBuilder()
                         .setName(product.name)
@@ -56,7 +79,14 @@ public class ProductController {
     }
 
     @PostMapping("/delete")
-    public ResponseEntity<Product> deleteProduct(@RequestBody DeleteProductRequest request) {
+    public ResponseEntity<Product> deleteProduct(@RequestHeader (value = "Authorization") String id,
+                                                 @RequestBody DeleteProductRequest request) throws IllegalAccessException{
+        String token = id.replace("Bearer ", "");
+        if (!jwtService.extractRole(token).equalsIgnoreCase("manager")
+                && !jwtService.extractRole(token).equalsIgnoreCase("admin")) {
+            throw new IllegalAccessException("You have no access.");
+        }
+
         Product savedProduct = productService.deleteProduct(request.UUID);
         return new ResponseEntity<>(savedProduct, HttpStatus.OK);
     }
